# docs/swagger.yaml
basePath: /atenea
definitions:
...
  models.ListaAlarma:
    properties:
      equipo:
        type: string
      id_equipo:
        type: integer
      msg_card_id:
        type: integer
      msg_id:
        type: integer
      msg_instance:
        type: integer
      msg_port:
        type: integer
      msg_set_time:
        type: string
      msg_severity:
        type: string
      msg_slot:
        type: integer
      msg_source_name:
        type: string
      msg_text:
        type: string
    type: object
...
host: IPServer:3021
info:
  contact:
    email: rgarciaxxx@yahoo.com.mx
    name: 'Soporte para las APIS: Raúl García'
    url: http://www.ineel.mx
  description: Contiene las APIs para atender las peticiones de la IHM hacia la base
    de datos ATENEA
  license:
    name: Atenea Inc.
    url: http://www.atenea.mx/LICENSE-2.0.html
  title: Servidor de APIs para ATENEA
  version: "1.0"
paths:
  /alarma/critica/equipo/{id}:
    get:
      description: Obtiene la lista de alarmas criticas correspondiente al identificador
        del equipo indicado
      parameters:
      - description: Identificador (id) del equipo
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.ListaAlarma'
            type: array
        "400":
          description: ""
        "404":
          description: ""
        "500":
          description: ""
      summary: Obtiene la lista de alarmas criticas de equipos
      tags:
      - alarma
  /alarma/critica/grupo/{id}:
    get:
      description: Obtiene la lista de alarmas criticas correspondiente al identificador
        del grupo indicado
      parameters:
      - description: Identificador (id) del grupo
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.ListaAlarma'
            type: array
        "400":
          description: ""
        "404":
          description: ""
        "500":
          description: ""
      summary: Obtiene la lista de alarmas criticas por grupos
      tags:
      - alarma
  /alarma/grupo/{id}:
    get:
      description: Obtiene la lista de alarmas correspondiente al identificador del
        grupo indicado
      parameters:
      - description: Identificador (id) del grupo
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.ListaAlarma'
            type: array
        "400":
          description: ""
        "404":
          description: ""
        "500":
          description: ""
      summary: Obtiene la lista de alarmas por grupos
      tags:
      - alarma
  /alarma/major/equipo/{id}:
    get:
      description: Obtiene la lista de alarmas major correspondiente al identificador
        del equipo indicado
      parameters:
      - description: Identificador (id) del equipo
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.ListaAlarma'
            type: array
        "400":
          description: ""
        "404":
          description: ""
        "500":
          description: ""
      summary: Obtiene la lista de alarmas major por equipos
      tags:
      - alarma
  /alarma/major/grupo/{id}:
    get:
      description: Obtiene la lista de alarmas major correspondiente al identificador
        del grupo indicado
      parameters:
      - description: Identificador (id) del grupo
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.ListaAlarma'
            type: array
        "400":
          description: ""
        "404":
          description: ""
        "500":
          description: ""
      summary: Obtiene la lista de alarmas major por grupos
      tags:
      - alarma
  /alarma/not-reachable/equipo/{id}:
    get:
      description: Obtiene la lista de alarmas not-reachable correspondiente al identificador
        del equipo indicado
      parameters:
      - description: Identificador (id) del equipo
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.ListaAlarma'
            type: array
        "400":
          description: ""
        "404":
          description: ""
        "500":
          description: ""
      summary: Obtiene la lista de alarmas not-reachable por equipos
      tags:
      - alarma
  /alarma/not-reachable/grupo/{id}:
    get:
      description: Obtiene la lista de alarmas not-reachable correspondiente al identificador
        del grupo indicado
      parameters:
      - description: Identificador (id) del grupo
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.ListaAlarma'
            type: array
        "400":
          description: ""
        "404":
          description: ""
        "500":
          description: ""
      summary: Obtiene la lista de alarmas not-reachable por grupos
      tags:
      - alarma
  /alarma/notify/equipo/{id}:
    get:
      description: Obtiene la lista de alarmas notify correspondiente al identificador
        del equipo indicado
      parameters:
      - description: Identificador (id) del equipo
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.ListaAlarma'
            type: array
        "400":
          description: ""
        "404":
          description: ""
        "500":
          description: ""
      summary: Obtiene la lista de alarmas notify por equipos
      tags:
      - alarma
  /alarma/notify/grupo/{id}:
    get:
      description: Obtiene la lista de alarmas notify correspondiente al identificador
        del grupo indicado
      parameters:
      - description: Identificador (id) del grupo
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.ListaAlarma'
            type: array
        "400":
          description: ""
        "404":
          description: ""
        "500":
          description: ""
      summary: Obtiene la lista de alarmas notify por grupos
      tags:
      - alarma
  /alarma/warning/equipo/{id}:
    get:
      description: Obtiene la lista de alarmas warning correspondiente al identificador
        del equipo indicado
      parameters:
      - description: Identificador (id) del equipo
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.ListaAlarma'
            type: array
        "400":
          description: ""
        "404":
          description: ""
        "500":
          description: ""
      summary: Obtiene la lista de alarmas warning por equipos
      tags:
      - alarma
  /alarma/warning/grupo/{id}:
    get:
      description: Obtiene la lista de alarmas warning correspondiente al identificador
        del grupo indicado
      parameters:
      - description: Identificador (id) del grupo
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.ListaAlarma'
            type: array
        "400":
          description: ""
        "404":
          description: ""
        "500":
          description: ""
      summary: Obtiene la lista de alarmas warning por grupos
      tags:
      - alarma
...
swagger: "2.0"